change PL Clocks(100MHz) and add network

+++ HLS/network/.autopilot/db/mnist_AXI_Stream.pragma.2.cpp
@@ -28713,7 +28713,7 @@ _ssdm_Unroll(0,0,0, "");
   if(i == array_length - 1){
    out.last = 1;
   }
-  out.data = array_head[i];
+  out.data = (int32_t)(array_head[i]);
   output_data << out;
  }
  return(0);
